Lua/APR 0.6-3 release and rockspec
After pushing the previous change and uploading lua-apr-0.6-2.zip I
tested LuaRocks which barfed about the ZIP archive (base directory
missing) so I tried to delete and re-upload but GitHub warned me that
this doesn't always work because of caching. That's why I've decided to
just bump the release number from 0.6-2 to 0.6-3.