Switch to Discount for Markdown conversion
markdown.lua was taking about ten seconds to convert the generated
documentation for the Lua/APR binding and this became annoying very
quickly, so I switched to Discount which runs in less than a second.

I checked the output and found a few errors (or maybe implementation
defined but undesirable behavior) so I fixed those as well -- the docs
can now be generated using either markdown.lua or Discount.