Support for XML parsing
This module is another contribution by zhiguo zhao. As with the DBD
binding I've changed quite a bit to integrate it with the Lua/APR
binding and added documentation and unit tests (85% test coverage).

Because I used a `feature branch' to integrate this module and worked in
this branch for a few days some other miscellaneous changes crept in:

 - apr.type() now supports thread objects and MD5/SHA1 contexts. While I
   was at it I also made all __tostring() implementations use the
   respective lua_apr_objtype structures so that every type string is
   only defined once in the source code.

 - Use LUA_APR_MSGSIZE define for error message buffer sizes.