Make child threads independant of their creators
Before this commit thread objects were allocated as userdata, which meant
that when the Lua state in which a thread was created got closed, the
userdata was also invalidated. This can cause crashes, for example when
the Lua interpreter exits while a background thread is still running.

The new code allocates thread objects with malloc() and uses reference
counting to decide which Lua state should free() the thread object when
neither Lua state will need it again.

When a thread that encountered an error is garbage collected it will print
the error message to stderr, unless the user has called thread:join() in
which case nothing is printed (this is a compromise between being
foolproof and not wanting to pollute stderr too much).