Bug fix: Give XML parser objects independent memory pools
The memory pool returned by to_pool() is used as a memory pool for
temporary allocations inside Lua/APR functions and is cleared on each
invocation of to_pool(). When it is cleared, child pools are automatically
destroyed. This is the reason that objects created by Lua/APR keep their
own private root pool which is only destroyed when the object is
garbage collected by the Lua garbage collector.

Evidently I forgot to change this before committing the XML module :-)