Support for HTTP request parsing using libapreq2
I've been planning a binding to libapreq2 for a while now but a recent
thread on lua-l* motivated me to dive in. Using etc/apreq_standalone.c
the HTTP request parsing module can be used independently from the rest
of the Lua/APR binding so that other projects can use the request
parsing without dragging in all of Lua/APR.

Note that the UNIX and Windows makefiles enable the libapreq2 binding by
default but you can disable it with "make DISABLE_APREQ=1". If and when
the dependency on libapreq2 turns out not to cause any problems I'll
probably switch it on permanently. I had some trouble getting libapreq2
to build on Windows initially, I've added instructions to NOTES.md.

* http://lua-users.org/lists/lua-l/2011-02/msg01443.html