Another buffer management bug fix!
I was working on a file I/O benchmark to see if the previous commit
might have impacted performance and found a bug in file:read('*n')
handling: If EOF was reached and there was no more buffered input,
file:read('*n') could try to convert the first byte in the buffer to a
number and if this succeeded, it would keep matching this value! Of
course nothing should be matched when there's no input to consume.