Bug fix for file:read(n) in text mode (+todos)
The buffered I/O interface would fail to convert a CR + LF sequence when CR
was the last buffered byte and the length of the buffered input was enough
to satisfy the read() call. This is now fixed by changing the logic and
checking for the corner case so it can be dealt with.

I've added some TODO items, I assume they're self explanatory :-)