Always join child threads, remove support for detachable threads
The thread queue tests would sometimes crash and it took me a while to
realize why. It turns out the tests didn't join the child thread and my
improvised "wait for all child threads to exit" code was buggy.

I then tried to detach the child thread in the test but got the same
crashes. In fact I couldn't get detached threads working without
having them crash every once in a while.

For this reason I've decided to remove the support for detached threads
and enforce joining of child threads: When a thread object is collected
the parent thread will automatically join the child thread.