Change how to tests are run (by calling them explicitly)
I recently released my Lua file type plug-in for Vim and added dynamic
completion which searches package.path and package.cpath and requires()
every available module to generate a list of valid symbols. When I tested
the completion it would hang for about 10 seconds and I didn't understand
why... Turns out it's not such a good idea to execute code when a module
is loaded: The completion code was executing the test suite!