Bug fix for apr.proc_fork() test (issue #5)
Oh my how embarrassing... The apr.proc_fork() test failed on the Debian
build bots for the hurd-i386 and mipsel architectures and because the
same test already worked on all other architectures I'd just assumed
until now that the problem wasn't with my code. Turns out it was!

The problem was that the test would first fork a child process and only
then generate a temporary file name used to communicate between the
parent and the forked child. I'm actually quite surprised that this test
worked on any platform at all! (IIRC random number generators are
supposed to be reinitialized after a fork()).

Anyway I'm guessing this is the reason that the apr.proc_fork() test
failed on hurd-i386 and mipsel so I'm hoping that this version of the
Lua/APR binding will be the first to successfully compile and test
on all platforms supported by Debian :-)