Bug fix for signal handling tests on kFreeBSD (issue #5)
The signal handling tests failed on kFreeBSD. Using a virtual machine
image from virtualboxes.org I figured out how to fix it. The problem
appears to be that os.execute() uses the shell to execute the external
command and on kFreeBSD the death of "grand children" is not reported
to the main process. By using apr.proc_create() and setting the command
type to program the tests pass on kFreeBSD and my own machine, so this
change appears to be an improvement :-)