Makefile improvements (don't fail when Markdown not installed)
While hunting for bugs in the signal handling module on a kFreeBSD
virtual machine I noticed that installation using LuaRocks would fail
when no Lua Markdown implementation was available, which is silly
because the LuaRocks ZIP archives include up to date documentation
so it really doesn't have to be regenerated. Anyway this should now
be fixed. I also moved all generation of HTML files to one place so
that I don't have to repeat the four etc/wrap.lua invocations.