LDAP connection handling module by zhiguo zhao (release 0.20)
Several months ago zhiguo zhao contributed an LDAP connection handling
module that combines the APR LDAP functions with LuaLDAP. Because I
didn't have any previous experience with LDAP it took me a while to get
familiar with LDAP and set up an LDAP server to be able to run tests.

The module I'm committing now contains work by the LuaLDAP authors,
zhiguo zhao and myself. I've updated the license in the README to
reflect this. According to my reading of the LuaLDAP license publishing
code based on LuaLDAP should be fine as long as the copyright notice
stays intact (it's in the README and the full license is also included
in src/ldap.c).

Right now this module can only search, but I'm planning to add functions
to add, modify, delete and rename LDAP directory entries. Also this
module is currently not tested on Windows because I can't get the LDAP
module to work at all: The code compiles but apr.ldap() fails with a
very non descriptive error.